extract permalink panel to own component
