+++ src/component/Permalink/Permalink.css
@@ -14,3 +14,8 @@
   border-radius: 2px;
   padding: 0 5px;
 }
+
+.permalink input:focus {
+  outline: none !important;
+  border: 1px solid var(--headerColor);
+}
